requirements.txt
pandas
requests
python-dotenv
pytrends
selenium